Added comments features, unlike option, merged texts & images into posts and removed follow user option

+++ src/utils/helper.js
@@ -4,7 +4,10 @@ import fs from 'fs';
 function delfile(filename){
     let filepath = path.join(process.cwd(), 'uploads', filename);
     //   process.cwd() = __dirname for es6
-    fs.unlinkSync(filepath)
+    if(fs.existsSync(filepath)) fs.unlinkSync(filepath);
 }
 
-export {delfile};
+const isImage = (file) => file?.mimetype?.startsWith('image/');
+const isVideo = (file) => file?.mimetype?.startsWith('video/');
+
+export {delfile, isImage, isVideo};
